Change num_urls_checked to bigint as it can easily overflow 2^31

+++ sql.d/schema/create_schema_tables.sql
@@ -458,7 +458,7 @@ CREATE TABLE statistics (
 	num_metapackages integer NOT NULL DEFAULT 0,
 	num_problems integer NOT NULL DEFAULT 0,
 	num_maintainers integer NOT NULL DEFAULT 0,
-	num_urls_checked integer NOT NULL DEFAULT 0,
+	num_urls_checked bigint NOT NULL DEFAULT 0,
 	num_reports_total integer NOT NULL DEFAULT 0,
 	num_reports_open integer NOT NULL DEFAULT 0
 );
